Solve medium find next right interval given a set of intervals using binary search and sorting
Leanings -- Implementation details of binary search. When the key is greater than the largest value prsent in the array, it returns - (size of array) - 1
Care must be taken to handle this, because converting this to positive number will return an index out of bounds.

Also if there are multiple occurrences of the key present in the array, any one of them can be returned. So need to manually traverse back and find the first occurrence.