Solve medium but hard find 132 pattern in array i.e find three numbers such that a[i] < a[k] < a[j] and i < j < k using monotonic stack
Learnings -- I solved this problem on my own, took a couple of hours but went through couple of different approaches and then finalized this. The crux of the problem was to fix i and then look for a valid k for each j. a[i] can be the smallest number on the left of a[j] and a[k] can be the largest number on the right smaller than a[j]. Once we find all three, check if they make 132 pattern.