Solve hard and tricky find the smallest substring that contains all the characters of the target string by sliding window.
Learnings -- In sliding window implementation, we are only interested in characters with +ve count. Increase the curCount only if the count of char[end] > 0. Decrease the curCount when count of char[begin] > 0. That's the trick. Incrementing/Decrementing count of characters that are not required wont have any affect because their count will always <= 0