Solve hard and tricky Dungeon Game, find the min health required at the start to reach the target provided the health cant be <= 0 at any state.
This is similar to BusRoutes where BFS needs to be applied but the condition for BFS has changed. Similarly here, the start is actually our target state, because we need the min health at the start. So at start node, we ask adjacent node on their min health required and update the health at start accordingly.