Solve hard classical jump game problem, find the min hops required to reach the last index via greedy and dp.
DP problem was easy to write, just update all the indexes reachable from current node to current jumps + 1

Greedy was difficult to visualize because I was not sure on when to update the jumps done so far. Trick was to visualize it as BFS problem instead and update the level count when end of current level is reached.