Solve hard and tricky evaluate the no of atoms in a chemical formula i.e ((H2)2Mg3)3, H=12, Mg=9
Trick was to use stack of map, start of each expression will contain the count of atoms of that expression, when a new expression is found '(', push a empty map upon stack.
When ')' is found, find the multiplier and multiply the current atom counts, merge the current atom counts with the previous expression.

This merge reduces the size of the result stack. Similar to DecodeString

Recursive solution is also possible but care must be taken to correctly recurse for the subexpression present under ()