Solve hard and tricky find the kth sorted permutation of the given input array using greedy + maths.
Backtracking solution is trivial and requires generating all k permutations of the input array.

Greedy solution fixes one digit at a time of the result; by figuring out how many permutations does a digit can generate if it's the first digit of the permutation. We recursively fix one digit and are left with n-1 digits.