Solve easy but tricky greedy problem, given the initial position of the chips, find the min cost to move them to the same position.
Min cost can be found if one of the initial state is chosen as the end state, so choose all of the initial position  as the target position and find the total cost to bring remaining chips to it's position.

This n^2 pattern seems to be pretty common. Can be done in linear time but ....