Solve medium and tricky problem, find no of moves required to make all the nodes in the tree with 1 coin.
Try to think about what info is required from the children, and update result and return value accordingly.
Here we need the surplus/deficit from left/right subtree. Once we know that, we can calculate the total surplus/deficit of current subtree.
Moves required at each node is the absolute value of surplus/deficit at that node e.g. if a node has surplus 3, it needs 3 moves to transfer the surplus because that node can have value = 1 in the end state.