Solve hard and tricky find min no of cameras to place to cover a binary tree. This is similar to vertex cover but is very tricky.
Trick is to consider it as a greedy problem rather than recursion problem, think about not placing a camera at the edge but at the center. So lead node should not have a camera, and instead it's parent should have a camera.

Sequence of covering would be covered (null pointers of leaf) -> uncovered -> place camera -> covered -> uncovered -> place camera ..