Solve medium and a bit tricky remove intervals that overlaps with a target interval.
Got confused during implementation trying to solve the question in log(n) using binary search as the input array is sorted. Checked the hint and realized that it can't be done in linear time because we have to create a new list and copy all the intervals in worst case.

So the simple solution is to check each interval and see if it overlaps with the target interval. If it does, split it.