Solve hard and tricky find the longest repeated substring in string using binary search + rolling hash
Binary search was not that difficult to identify but rolling hash took a lot of time due to modulo arithmetic in handling negative hashes. Also care must be taken to properly calculate new hash. Take care of indices and the multiplication factor.

LCS can also be used to find the answer but it would have been n^2 and would have timed out due to higher input bound