Solve medium and a bit tricky problem, find the total probability of knight remaining on the chessboard after making k moves.
I am a bit surprised to do this in first attempt. Think I am getting a little good :P

A gotcha to make the solution memoizable is to return results of each subproblem independently of another. Initially I was thinking of passing in the current prob as a prefix multiplier but then discarded it