Solve medium and extremely tricky problem, find the path with maximum minimum value possible from source to target in a grid using  3 approaches: binary search or priority queue or union find
Learnings -- In case of grid problems where traversal in 4 direction is permissible make sure to handle cycles. If you are memoizing the solution, make sure the memoized value does not depend upon the initial path that was taken to visit that cell.

Binary search is to go to approach for dealing with maximum minimum problems