Solve medium find whether there is a increasing subsequence of size 3 in linear time and constant space.
In order to solve it using linear space, I kept track of smallest number on the left and largest number on the right and compared each number.
In order to solve in constant space, just keep track of two smallest numbers. If you find a number greater than second largest number, we got a valid triplet.