Update hard and extremely trick find the maximum of minimum sweetness that can satisfy k+1 persons using binary search
I was not able to understand the logic of finding how many persons can be satisfied by choosing a specific min sweetness. The trick was to increase the persons satisfied count only if curSweetness >= target sweetness. This way if there is some trailing subarray that can't satisfy a person won't get included in the overall count.