Solve medium and extremely tricky find the max no of moves required to make sure no two stones share the same row or same column by DFS and union find
Trick to solve the problem was to find no of connected components and just keep only one stone of each component. So answer would be total stones - no of components

Learnings -- 1. Sometimes 2D points can be visualized as 0,1 in a 2D grid, which can be expanded to a graph problem if we can visualize the edges correctly. Here two points are connected if they are on the same row or same column.
2. Union find on 2D points is a bit tricky to visualize. Here we need to union row and col of current index so that in future if we get a new index with the same row and different col, we can still make it part of the same component.