Solve medium and tricky find single element in sorted array such that all other elements repeat twice using binary search.
Trick was to note that numbers on the left side of the target have even and odd indices with same value whereas on the right, odd and even indices have the same value. Use this fact to reduce this to T,T,T,F,F,F,F along with handling edge cases when target is the first or last element of the array.
While solving binary search cases, consider each index as mid and try to solve it.