Solve hard and extremely tricky find critical connections in a network using tarzan's algorithm for detecting back edges.
Keep track of the time a node was first visited, if it's neighbourTime > firstVisitedTime, we have a critical connection. Meanwhile if a node is connected to a node with smaller time, this means there is another path to this node hence current path isn't critical. Keep updating the time of an index by taking the min of neighbourTime and curNodeTime