Solve medium and extremely tricky reconstruction problem, reconstruct original digits from its scrambled english representation using tricky observation.
Initially I tried to solve using greedy where I greedily checked whether I could assign "zero", "one" or so on.. but it didn't work for input like "twonine", because "one" is present in the input and it didn't result in proper utilization of characters.

Trick was to find the unique characters of each digit and try to leverage that uniqueness to find how many times that digit exists e.g. 'z' is unique to "zero" and so on. Also check must be done in an ordered manner.