Solve hard and extremely tricky design LFU cache using hashmap of key -> node + Doubly linked list of frequency -> set of nodes with the respective frequency.
I was not able to fully solve this problem on my own. I was not able to cleanly implement how to find the smallest frequency and reshuffle nodes to +1 frequency node in DLL.
Once I saw the paper, I was able to visualize it clearly and was able to implement it cleanly as well.