Add another solutions to hard and extremely tricky problem, find the shortest path in a grid with obstacle eliminations using Djikstra and BFS by maintaining additional state per node.
Learnings: Since this was a unweighted graph, BFS was the way to go. But we need to maintain additional state per node, so either maintain a nested array with additional dimension or maintain additional array similar to cost array. Visit an edge only if it reduces the cost or reduces no of eliminations. Last but not the least. Maintain a consistent state every node.